+++ src/proto_018_Proxford/lib_sc_rollup_layer2/sc_rollup_services.ml
@@ -333,6 +333,15 @@ module Block = struct
       ~output:Data_encoding.(list string)
       (path / "durable" / Sc_rollup.Kind.to_string pvm_kind / "subkeys")
 
+  let durable_state_values (pvm_kind : Sc_rollup.Kind.t) =
+    Tezos_rpc.Service.get_service
+      ~description:
+        "Retrieve values directly under a given key from PVM durable storage. \
+         PVM state is taken with respect to the specified block level."
+      ~query:Query.key_query
+      ~output:Data_encoding.(list (obj2 (req "key" string) (req "value" bytes)))
+      (path / "durable" / Sc_rollup.Kind.to_string pvm_kind / "values")
+
   let status =
     Tezos_rpc.Service.get_service
       ~description:"PVM status at block"